RIP-116, retire IFrameMixin in favor of global utils
