+++ package.json
@@ -41,7 +41,7 @@
     "eslint": "^8.36.0",
     "eslint-plugin-vue": "^9.9.0",
     "sass": "^1.59.3",
-    "typescript": "^5.0.2",
+    "typescript": "^4.9.5",
     "vite": "^4.2.1",
     "vite-plugin-vuetify": "^1.0.2",
     "vue-tsc": "^1.2.0"
