+++ src/main.ts
@@ -3,7 +3,7 @@ import App from './App.vue'
 import axios from 'axios'
 import moment from 'moment'
 import {createApp} from 'vue'
-import {initializeAxios} from './utils'
+import {initializeAxios, isInIframe} from './utils'
 import {registerPlugins} from '@/plugins'
 import {useContextStore} from '@/stores/context'
 import router from '@/router'
@@ -17,6 +17,7 @@ initializeAxios(app, axios)
 
 // Globals
 app.config.globalProperties.$_ = _
+app.config.globalProperties.$isInIframe = isInIframe
 app.config.globalProperties.$moment = moment
 app.config.globalProperties.$ready = (label: string, focusTarget?: string) => useContextStore().loadingComplete(label, focusTarget)
